backend/pyproject.toml
[project]
name = "oncomind-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.12.15",
    "aiosignal>=1.4.0",
    "annotated-types>=0.7.0",
    "anthropic>=0.67.0",
    "anyio>=4.10.0",
    "arxiv>=2.2.0",
    "asyncpg>=0.30.0",
    "attrs>=25.3.0",
    "backoff>=2.2.1",
    "beautifulsoup4>=4.13.5",
    "biopython>=1.85",
    "boto3>=1.40.30",
    "botocore>=1.40.30",
    "cachetools>=5.5.2",
    "certifi>=2025.8.3",
    "cffi>=1.17.1",
    "chardet>=5.2.0",
    "charset-normalizer>=3.4.3",
    "click>=8.2.1",
    "cloud-sql-python-connector>=1.18.4",
    "colorama>=0.4.6",
    "cryptography>=45.0.7",
    "curl-cffi>=0.13.0",
    "dataclasses-json>=0.6.7",
    "deepdiff>=8.6.1",
    "distro>=1.9.0",
    "docstring-parser>=0.17.0",
    "duckduckgo-search>=8.1.1",
    "email-validator>=2.3.0",
    "emoji>=2.14.1",
    "fastapi>=0.116.1",
    "fastapi-mail>=1.5.0",
    "feedparser>=6.0.12",
    "filelock>=3.19.1",
    "filetype>=1.2.0",
    "firecrawl>=4.3.6",
    "fireworks-ai>=0.17.4",
    "frozenlist>=1.7.0",
    "fsspec>=2025.9.0",
    "google-ai-generativelanguage>=0.6.6",
    "google-api-core>=2.25.1",
    "google-api-python-client>=2.181.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-cloud-aiplatform>=1.113.0",
    "google-cloud-bigquery>=3.37.0",
    "google-cloud-core>=2.4.3",
    "google-cloud-resource-manager>=1.14.2",
    "google-cloud-storage>=2.19.0",
    "google-crc32c>=1.7.1",
    "google-generativeai>=0.7.2",
    "google-resumable-media>=2.7.2",
    "googleapis-common-protos>=1.70.0",
    "greenlet>=3.2.4",
    "grpc-google-iam-v1>=0.14.2",
    "grpcio>=1.74.0",
    "grpcio-status>=1.62.3",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httplib2>=0.31.0",
    "httpx>=0.27.2",
    "httpx-sse>=0.4.1",
    "huggingface-hub>=0.34.4",
    "idna>=3.10",
    "jinja2>=3.1.6",
    "jiter>=0.11.0",
    "jmespath>=1.0.1",
    "joblib>=1.5.2",
    "jsonpatch>=1.33",
    "jsonpath-python>=1.0.6",
    "jsonpointer>=3.0.0",
    "jupyter>=1.1.1",
    "kay>=0.1.2",
    "langchain>=0.2.17",
    "langchain-community>=0.2.19",
    "langchain-core>=0.2.43",
    "langchain-google-genai>=1.0.10",
    "langchain-google-vertexai>=1.0.10",
    "langchain-openai>=0.1.25",
    "langchain-robocorp>=0.0.10.post1",
    "langchain-text-splitters>=0.2.4",
    "langdetect>=1.0.9",
    "langgraph>=0.4.5",
    "langgraph-checkpoint>=2.1.1",
    "langgraph-prebuilt>=0.1.8",
    "langgraph-sdk>=0.2.6",
    "langserve>=0.2.3",
    "langsmith>=0.1.147",
    "lxml>=6.0.1",
    "marshmallow>=3.26.1",
    "multidict>=6.6.4",
    "mypy-extensions>=1.1.0",
    "nest-asyncio>=1.6.0",
    "nltk>=3.9.1",
    "notebook>=7.4.5",
    "numpy>=1.26.4",
    "openai>=1.107.2",
    "orderly-set>=5.5.0",
    "orjson>=3.11.3",
    "ormsgpack>=1.10.0",
    "packaging>=24.2",
    "passlib[bcrypt]>=1.7.4",
    "pdfminer-six>=20250506",
    "pg8000>=1.31.4",
    "pgvector>=0.4.1",
    "pillow>=11.3.0",
    "praw>=7.8.1",
    "prawcore>=2.4.0",
    "propcache>=0.3.2",
    "proto-plus>=1.26.1",
    "protobuf>=4.25.8",
    "psycopg2-binary>=2.9.10",
    "pyasn1>=0.6.1",
    "pyasn1-modules>=0.4.2",
    "pycparser>=2.22",
    "pydantic>=2.11.7",
    "pydantic-core>=2.33.2",
    "pydantic-settings>=2.10.1",
    "pyparsing>=3.2.4",
    "pypdf>=6.0.0",
    "pypdf2>=3.0.1",
    "python-dateutil>=2.9.0.post0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-iso639>=2025.2.18",
    "python-jose[cryptography]>=3.5.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "rapidfuzz>=3.14.1",
    "regex>=2025.9.1",
    "requests>=2.32.5",
    "requests-toolbelt>=1.0.0",
    "rsa>=4.9.1",
    "s3transfer>=0.14.0",
    "setuptools>=80.9.0",
    "sgmllib3k>=1.0.0",
    "shapely>=2.1.1",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "socksio>=1.0.0",
    "soupsieve>=2.8",
    "sqlalchemy>=2.0.43",
    "sse-starlette>=3.0.2",
    "starlette>=0.47.2",
    "streamlit>=1.48.1",
    "tabulate>=0.9.0",
    "tavily-python>=0.7.12",
    "tenacity>=8.5.0",
    "tiktoken>=0.11.0",
    "tokenizers>=0.22.0",
    "tomli-w>=1.2.0",
    "tqdm>=4.67.1",
    "types-protobuf>=6.30.2.20250914",
    "types-requests>=2.32.4.20250913",
    "typing-extensions>=4.14.1",
    "typing-inspect>=0.9.0",
    "typing-inspection>=0.4.1",
    "unstructured>=0.18.14",
    "unstructured-client>=0.42.3",
    "update-checker>=0.18.0",
    "uritemplate>=4.2.0",
    "urllib3>=2.5.0",
    "uvicorn>=0.35.0",
    "websocket-client>=1.8.0",
    "websockets>=14.2",
    "wikipedia>=1.4.0",
    "wrapt>=1.17.3",
    "xmltodict>=1.0.0",
    "xxhash>=3.5.0",
    "yarl>=1.20.1",
    "zstandard>=0.25.0",
]
